docs: Include files are always searched at standard location, too.
This removes the need to specify the location where yml2 was unpacked or
installed as YML_PATH. So this now works:

   YML_PATH=/foo ./yml2proc --xml2yml

The standard system location <prefix>/share/yml2 is now always appended to
the include search path. If this path does not exist, the directory one
above the `backend.py` file is added, allowing the same behaviour for
uninstalled yml2.

In the Makefile enforce the .ysl2 files to be searched in the current
directory first.